Extract AtlasMap WebView in a dedicated class #36
